requirements.txt
asgiref==3.5.1
Django~=3.2.3
djangorestframework~=3.12.4
install==1.3.5
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2

djangorestframework-jwt~=1.11.0